RYSK-4: Separate out logic for readability.
